[refactor] - chat context -> customhook으로 변경
