+++ nbc-nep-front/src/components/metaverse/metaverseChat/MetaverseChat.tsx
@@ -3,8 +3,6 @@ import MetaverseChatForm from "@/components/metaverse/metaverseChat/metaverseCha
 import MetaverseChatList from "@/components/metaverse/metaverseChat/metaverseChatBar/MetaverseChatList";
 import { dmChatAlarmState } from "@/components/metaverse/types/ChatAlarmType";
 import { DMListCard } from "@/components/metaverse/types/metaverse";
-import { MetaverseChatProvider } from "@/context/MetaverseChatProvider";
-// import { usePlayerContext } from "@/context/MetaversePlayerProvider";
 import useChatAlarm from "@/hooks/GNB/useChatAlarm";
 import useMetaversePlayer from "@/hooks/metaverse/useMetaversePlayer";
 import { useGetLastDMList } from "@/hooks/query/useSupabase";
@@ -16,7 +14,6 @@ import styled from "styled-components";
 
 export default function MetaverseChat() {
   const { isOpenChat, chatType } = useAppSelector((state) => state.chatType);
-  // const { id, spaceId } = usePlayerContext();
   const { id, spaceId } = useMetaversePlayer();
   const { handleSetDmChatAlarmState } = useChatAlarm();
   const dmList = useGetLastDMList(spaceId, id);
@@ -52,18 +49,16 @@ export default function MetaverseChat() {
   }, [dmList]);
 
   return (
-    <MetaverseChatProvider>
-      <StMetaverseGlobalChatWrapper $isOpenChat={isOpenChat}>
-        {chatType === "GLOBAL" ? (
-          <>
-            <MetaverseChatList />
-            <MetaverseChatForm />
-          </>
-        ) : (
-          <MetaverseDmList dmList={dmList} />
-        )}
-      </StMetaverseGlobalChatWrapper>
-    </MetaverseChatProvider>
+    <StMetaverseGlobalChatWrapper $isOpenChat={isOpenChat}>
+      {chatType === "GLOBAL" ? (
+        <>
+          <MetaverseChatList />
+          <MetaverseChatForm />
+        </>
+      ) : (
+        <MetaverseDmList dmList={dmList} />
+      )}
+    </StMetaverseGlobalChatWrapper>
   );
 }
